+++ tests/fixtures/fs.py
@@ -1,3 +1,4 @@
+import os
 from typing import Any, Generator
 
 import pytest
@@ -13,3 +14,8 @@ def tmpfs(
         return tmp_path_factory.mktemp(basename)
 
     yield _tmp
+
+
+def get_testdata_file_path(file: str) -> str:
+    testdata_dir = os.path.join(os.path.dirname(__file__), "..", "testdata")
+    return os.path.join(testdata_dir, file)
