+++ tests/fixtures/metrics.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 
 from fixtures import types
-from fixtures.utils import parse_timestamp
+from fixtures.time import parse_timestamp
 
 
 class MetricsTimeSeries(ABC):
